# PERF: Special-case single-converter in loadtxt.