~5-13% speedup: `[*map(conv, items)]` (single converter, which is quite
common) is much faster than `[conv(val) for conv, val in zip(converters,
vals)]`.

`_loadtxt_floatconv` and `fencode` were lifted out so that every
"instance" of them is the same, allowing checking for whether there's
different converters in use (actually, it looks like two `floatconv`s
returned by two separate calls to `_getconv` have the same identity, but
we don't need to rely on that.

```
       before           after         ratio
     [cc7f1504]       [3a1e0a57]
     <main>           <loadtxtsingleconverter>
-      31.1±0.3μs       29.5±0.1μs     0.95  bench_io.LoadtxtCSVdtypes.time_loadtxt_dtypes_csv('object', 10)
-      9.17±0.1ms       8.59±0.2ms     0.94  bench_io.LoadtxtUseColsCSV.time_loadtxt_usecols_csv([1, 3])
-         236±3μs          221±2μs     0.94  bench_io.LoadtxtCSVdtypes.time_loadtxt_dtypes_csv('complex128', 100)
-         224±6μs          208±5μs     0.93  bench_io.LoadtxtCSVdtypes.time_loadtxt_dtypes_csv('int64', 100)
-       193±0.7μs        178±0.5μs     0.92  bench_io.LoadtxtCSVdtypes.time_loadtxt_dtypes_csv('int32', 100)
-         219±8ms          202±4ms     0.92  bench_io.LoadtxtCSVdtypes.time_loadtxt_dtypes_csv('int64', 100000)
-      7.16±0.2ms      6.61±0.09ms     0.92  bench_io.LoadtxtUseColsCSV.time_loadtxt_usecols_csv(2)
-         188±3ms          173±2ms     0.92  bench_io.LoadtxtCSVdtypes.time_loadtxt_dtypes_csv('int32', 100000)
-         250±4ms        229±0.7ms     0.92  bench_io.LoadtxtCSVSkipRows.time_skiprows_csv(0)
-         181±1μs          166±2μs     0.92  bench_io.LoadtxtCSVdtypes.time_loadtxt_dtypes_csv('object', 100)
-      18.2±0.3ms      16.6±0.08ms     0.92  bench_io.LoadtxtCSVdtypes.time_loadtxt_dtypes_csv('int32', 10000)
-       182±0.7μs          165±2μs     0.91  bench_io.LoadtxtCSVComments.time_comment_loadtxt_csv(100)
-      11.3±0.1ms      10.2±0.07ms     0.91  bench_io.LoadtxtUseColsCSV.time_loadtxt_usecols_csv([1, 3, 5, 7])
-       183±0.5μs          166±1μs     0.90  bench_io.LoadtxtCSVdtypes.time_loadtxt_dtypes_csv('float64', 100)
-      17.2±0.1ms      15.5±0.07ms     0.90  bench_io.LoadtxtCSVComments.time_comment_loadtxt_csv(10000)
-      12.7±0.4ms      11.4±0.09ms     0.90  bench_io.LoadtxtReadUint64Integers.time_read_uint64(10000)
-     1.28±0.02ms      1.15±0.01ms     0.90  bench_io.LoadtxtReadUint64Integers.time_read_uint64(1000)
-      12.9±0.4ms      11.5±0.09ms     0.90  bench_io.LoadtxtReadUint64Integers.time_read_uint64_neg_values(10000)
-        717±20μs          640±4μs     0.89  bench_io.LoadtxtReadUint64Integers.time_read_uint64_neg_values(550)
-         183±3ms          163±3ms     0.89  bench_io.LoadtxtCSVComments.time_comment_loadtxt_csv(100000)
-         182±3ms        162±0.6ms     0.89  bench_io.LoadtxtCSVdtypes.time_loadtxt_dtypes_csv('float64', 100000)
-      17.5±0.1ms       15.6±0.1ms     0.89  bench_io.LoadtxtCSVdtypes.time_loadtxt_dtypes_csv('object', 10000)
-        726±30μs          643±2μs     0.89  bench_io.LoadtxtReadUint64Integers.time_read_uint64(550)
-         182±3ms          161±2ms     0.88  bench_io.LoadtxtCSVdtypes.time_loadtxt_dtypes_csv('float32', 100000)
-      17.2±0.1ms       15.2±0.1ms     0.88  bench_io.LoadtxtCSVdtypes.time_loadtxt_dtypes_csv('float32', 10000)
-      17.3±0.1ms       15.2±0.1ms     0.88  bench_io.LoadtxtCSVdtypes.time_loadtxt_dtypes_csv('float64', 10000)
-        20.4±1ms      17.8±0.04ms     0.87  bench_io.LoadtxtCSVdtypes.time_loadtxt_dtypes_csv('str', 10000)
```