@anntzer 2021-08-06
```
$ python runtests.py --bench-compare HEAD~ Loadtxt
<elided>

       before           after         ratio
     [88776607]       [c42f57b5]
     <main>           <loadtxtsingleconverter>
+      44.8±0.5ms       47.1±0.3ms     1.05  bench_io.LoadtxtCSVDateTime.time_loadtxt_csv_datetime(20000)
-      9.15±0.1ms       8.71±0.1ms     0.95  bench_io.LoadtxtUseColsCSV.time_loadtxt_usecols_csv([1, 3])
-         694±5μs          661±9μs     0.95  bench_io.LoadtxtReadUint64Integers.time_read_uint64(550)
-     12.4±0.02ms       11.8±0.2ms     0.95  bench_io.LoadtxtReadUint64Integers.time_read_uint64(10000)
-        1.25±0ms      1.19±0.01ms     0.95  bench_io.LoadtxtReadUint64Integers.time_read_uint64_neg_values(1000)
-       246±0.9ms          234±1ms     0.95  bench_io.LoadtxtCSVSkipRows.time_skiprows_csv(0)
-      19.3±0.2ms       18.3±0.2ms     0.95  bench_io.LoadtxtCSVdtypes.time_loadtxt_dtypes_csv('str', 10000)
-     7.03±0.07ms      6.66±0.06ms     0.95  bench_io.LoadtxtUseColsCSV.time_loadtxt_usecols_csv(2)
-         203±1μs          192±3μs     0.94  bench_io.LoadtxtCSVdtypes.time_loadtxt_dtypes_csv('str', 100)
-         246±1ms          231±2ms     0.94  bench_io.LoadtxtCSVSkipRows.time_skiprows_csv(500)
-         224±2ms          210±2ms     0.94  bench_io.LoadtxtCSVSkipRows.time_skiprows_csv(10000)
-         228±1ms          214±1ms     0.94  bench_io.LoadtxtCSVdtypes.time_loadtxt_dtypes_csv('complex128', 100000)
-         196±2ms          184±2ms     0.94  bench_io.LoadtxtCSVdtypes.time_loadtxt_dtypes_csv('str', 100000)
-         178±2ms        167±0.5ms     0.94  bench_io.LoadtxtCSVComments.time_comment_loadtxt_csv(100000)
-         185±1ms          174±1ms     0.94  bench_io.LoadtxtCSVdtypes.time_loadtxt_dtypes_csv('int32', 100000)
-     17.9±0.05ms       16.7±0.1ms     0.94  bench_io.LoadtxtCSVdtypes.time_loadtxt_dtypes_csv('int32', 10000)
-         178±2ms        167±0.3ms     0.94  bench_io.LoadtxtCSVdtypes.time_loadtxt_dtypes_csv('float32', 100000)
-         235±1μs          220±2μs     0.94  bench_io.LoadtxtCSVdtypes.time_loadtxt_dtypes_csv('complex128', 100)
-      17.1±0.1ms      16.0±0.07ms     0.93  bench_io.LoadtxtCSVdtypes.time_loadtxt_dtypes_csv('object', 10000)
-      22.1±0.2ms      20.6±0.08ms     0.93  bench_io.LoadtxtCSVdtypes.time_loadtxt_dtypes_csv('complex128', 10000)
-      17.1±0.1ms      15.9±0.07ms     0.93  bench_io.LoadtxtCSVComments.time_comment_loadtxt_csv(10000)
-       181±0.9μs          169±1μs     0.93  bench_io.LoadtxtCSVComments.time_comment_loadtxt_csv(100)
-      11.2±0.1ms      10.4±0.07ms     0.93  bench_io.LoadtxtUseColsCSV.time_loadtxt_usecols_csv([1, 3, 5, 7])
-      17.0±0.3ms       15.9±0.1ms     0.93  bench_io.LoadtxtCSVdtypes.time_loadtxt_dtypes_csv('float32', 10000)
-         180±2ms        167±0.5ms     0.93  bench_io.LoadtxtCSVdtypes.time_loadtxt_dtypes_csv('float64', 100000)
-       183±0.4μs        170±0.6μs     0.93  bench_io.LoadtxtCSVdtypes.time_loadtxt_dtypes_csv('float64', 100)
-       180±0.9μs        167±0.8μs     0.93  bench_io.LoadtxtCSVdtypes.time_loadtxt_dtypes_csv('object', 100)
-       193±0.6μs        179±0.5μs     0.93  bench_io.LoadtxtCSVdtypes.time_loadtxt_dtypes_csv('int32', 100)
-         184±1μs          170±1μs     0.93  bench_io.LoadtxtCSVdtypes.time_loadtxt_dtypes_csv('float32', 100)
-      17.1±0.2ms      15.9±0.04ms     0.93  bench_io.LoadtxtCSVdtypes.time_loadtxt_dtypes_csv('float64', 10000)
-         182±3ms          168±1ms     0.92  bench_io.LoadtxtCSVdtypes.time_loadtxt_dtypes_csv('object', 100000)
-         219±5ms          201±2ms     0.92  bench_io.LoadtxtCSVdtypes.time_loadtxt_dtypes_csv('int64', 100000)
-      21.0±0.6ms       19.3±0.1ms     0.92  bench_io.LoadtxtCSVdtypes.time_loadtxt_dtypes_csv('int64', 10000)
-         225±5μs        203±0.8μs     0.90  bench_io.LoadtxtCSVdtypes.time_loadtxt_dtypes_csv('int64', 100)
```